Fix auto-placement with more complex layouts
The previous code used `document.body.clientHeight` to determine
if an element was out of the viewport. Unfortunately, this can cause
problems with layouts where the body element height is set to 100%.

Instead, we move back to using window.innerHeight, but add the current
scrollY, to determine if the bottom edge is off the screen. This way,
we can dynamically tell if the tooltip element would be scrolled out
of the user's view, and position it appropriately.